Border those blockquotes with orange
We already had this in place, but the white chalky line isn't visible on the white background. Fortunately, there is also already an orange version. If I had any idea how to run this css through less, maybe I could try it on dev. But I'm timid and uncertain. Catherine please review.