Width issues … oy! Need to make sure there is always a bit of margin left and right. They have internal complaints about the text extending all the way to the browser window’s edge.(force margin on <1240)
I’m still seeing the cut-off “Let’s Talk” button at times. But I do have something more specific to go on. The “Let’s talk” image itself is too big when I see this. So, it’s not always scaling to fit the space assigned to it. And it’s even worse at tablet width. I’m only getting it on the laptop and not the desktop. Can’t figure that out, because they’re both running the latest Chrome on El Capitan, and it doesn’t seem to change when I switch screen res.

Add size auto